requirements_rheed.txt
PyQt5>=5.15
opencv-python>=4.5
matplotlib>=3.5
numpy>=1.21
pyyaml>=6.0
